fix: diff docker tool - various fixes

+++ dev/diff/cli.js
@@ -286,18 +286,20 @@ async function main () {
                 ext: 'tgz'
               })
 
-              // Add missing files not present in release tarball
-              task.title = `Add missing diff tool files...`
-              await fs.ensureDir(path.join(config.target, 'dev/diff'))
-              await fs.copy(path.join(config.source, 'dev/diff/prepare.sh'), path.join(config.target, 'dev/diff/prepare.sh'))
-              await fs.copy(path.join(config.source, 'dev/diff/settings_local.py'), path.join(config.target, 'dev/diff/settings_local.py'))
-
               task.title = `Fetched latest release to ${config.target}`
               break
             }
             default: {
               throw new Error('Invalid selection. Exiting...')
             }
+          }
+
+          // Add missing files not present in branch
+          if (!(await fs.pathExists(path.join(config.target, 'dev/diff')))) {
+            task.output = `Add missing diff tool files...`
+            await fs.ensureDir(path.join(config.target, 'dev/diff'))
+            await fs.copy(path.join(config.source, 'dev/diff/prepare.sh'), path.join(config.target, 'dev/diff/prepare.sh'))
+            await fs.copy(path.join(config.source, 'dev/diff/settings_local.py'), path.join(config.target, 'dev/diff/settings_local.py'))
           }
         }
       },
@@ -496,28 +498,28 @@ async function main () {
           {
             title: 'Preparing source Datatracker instance...',
             task: async (subctx, subtask) => {
-              await executeCommand (subtask, containers.appSource, ['bash', '-c', 'chmod +x ./dev/diff/prepare.sh'])
-              await executeCommand (subtask, containers.appSource, ['bash', './dev/diff/prepare.sh'])
+              await executeCommand(subtask, containers.appSource, ['bash', '-c', 'chmod +x ./dev/diff/prepare.sh'])
+              await executeCommand(subtask, containers.appSource, ['bash', './dev/diff/prepare.sh'])
               subtask.title = `Preparing source Datatracker instance - Running checks...`
-              await executeCommand (subtask, containers.appSource, ['bash', '-c', './ietf/manage.py check'])
+              await executeCommand(subtask, containers.appSource, ['bash', '-c', './ietf/manage.py check'])
               subtask.title = `Preparing source Datatracker instance - Applying migrations...`
-              await executeCommand (subtask, containers.appSource, ['bash', '-c', './ietf/manage.py migrate'])
+              await executeCommand(subtask, containers.appSource, ['bash', '-c', './ietf/manage.py migrate'])
               subtask.title = `Source Datatracker instance is now ready.`
             }
           },
           {
             title: 'Preparing target Datatracker instance...',
             task: async (subctx, subtask) => {
-              await executeCommand (subtask, containers.appTarget, ['bash', '-c', 'chmod +x ./dev/diff/prepare.sh'])
-              await executeCommand (subtask, containers.appTarget, ['bash', './dev/diff/prepare.sh'])
+              await executeCommand(subtask, containers.appTarget, ['bash', '-c', 'chmod +x ./dev/diff/prepare.sh'])
+              await executeCommand(subtask, containers.appTarget, ['bash', './dev/diff/prepare.sh'])
               subtask.title = `Run target Datatracker instance - Running checks...`
-              await executeCommand (subtask, containers.appTarget, ['bash', '-c', './ietf/manage.py check'])
+              await executeCommand(subtask, containers.appTarget, ['bash', '-c', './ietf/manage.py check'])
               subtask.title = `Run target Datatracker instance - Applying migrations...`
-              await executeCommand (subtask, containers.appTarget, ['bash', '-c', './ietf/manage.py migrate'])
+              await executeCommand(subtask, containers.appTarget, ['bash', '-c', './ietf/manage.py migrate'])
               subtask.title = `Run target Datatracker instance - Starting server...`
-              executeCommand (subtask, containers.appTarget, ['bash', '-c', './ietf/manage.py runserver 0.0.0.0:8000 --settings=settings_local'])
+              executeCommand(subtask, containers.appTarget, ['bash', '-c', './ietf/manage.py runserver 0.0.0.0:8000 --settings=settings_local'])
               subtask.title = `Run target Datatracker instance - Waiting for server to accept connections...`
-              await executeCommand (subtask, containers.appTarget, ['bash', '-c', '/usr/local/bin/wait-for localhost:8000 -t 120'])
+              await executeCommand(subtask, containers.appTarget, ['bash', '-c', '/usr/local/bin/wait-for localhost:8000 -t 300'])
               subtask.title = `Target Datatracker instance is now ready and accepting connections.`
             }
           }
@@ -534,6 +536,7 @@ async function main () {
       {
         title: 'Run crawl tool',
         task: async (ctx, task) => {
+          task.output = 'Starting ./bin/test-crawl...'
           diffOutput = await executeCommand (task, containers.appTarget, ['bash', '-c', `./bin/test-crawl --settings=ietf.settings_testcrawl --diff http://dt-diff-app-target:8000/ ${config.options.join(' ')}`], true)
         }
       }
